Fixed bug showing all cards in one card field. A separate component has been created for this
