+++ src/components/AI/Cards/Cards bug.tsx
